Added a basic shell script for testing, and some examples
Files to be tested are in the folder clam/tests, and each has a corresponding
"result file" ending in .res -- so "add" goes with "add.res" etc.
tester.sh just goes through all the test files, tries to compile and run them,
and reports any problems - i.e. failure to compile, or unexpected output etc.

(The examples are for the calculator and not actual CLAM, so I expect we'll scrap them later.)