Massive commit which adds Syntax error handling
This also beefs up the Makefile to auto-calculate dependencies on the
fly for link and compile time. All you have to do is specify which
modules to compile :-)

I have brought in a few of the "ExtLib" files from here:
    http://code.google.com/p/ocaml-extlib/
It's a nice library, but I didn't want to force everyone to install it
on their machine.

ISSUE: autodependency calculation doesn't work for incremental builds.
WORK-AROUND: make clean && make
(will be fixed at some point in the future)